Fix compilation for Snow Leopard. Patch by Argyris Kyrtzidis!
git-svn-id: https://llvm.org/svn/llvm-project/vmkit/trunk@82035 91177308-0d34-0410-b5e6-96231b3b80d8

+++ lib/Mvm/Allocator/gcmapper.cpp
@@ -34,7 +34,7 @@ static inline void* manual_mremap(void * addr, int old_size, int new_size, int f
 {
   void * res = addr;
   if (new_size < old_size)
-    DO_MUNMAP((void*)((int)addr + new_size), old_size - new_size);
+    DO_MUNMAP((void*)((intptr_t)addr + new_size), old_size - new_size);
   else if (new_size > old_size)
     // Use of MAP_FIXED is discouraged...
     // res = mmap(addr, new_size, PROT_READ | PROT_WRITE, MAP_PRIVATE | MAP_ANON, -1, MAP_FIXED);
